Prepare ops.bb recipe to move repo to cmake

+++ yocto/openswitch/meta-distro-openswitch/recipes-ops/ops/ops.bb
@@ -14,10 +14,26 @@ PV = "git${SRCPV}"
 
 S = "${WORKDIR}/git"
 
+# XXX: Remove this when the repo moves to cmake
+B = "${S}"
+
 FILES_${PN} = "/usr/share/openvswitch/ /usr/share/openvswitch/*.extschema /usr/share/openvswitch/*.xml /usr/share/openvswitch/*.ovsschema"
 
 OPS_SCHEMA_PATH="${S}/schema"
 
+inherit cmake
+
+# XXX: Remove this when the repo moves to cmake
+do_configure() {
+	if test -e CMakeLists.txt ; then
+		cmake_do_configure
+	fi
+}
+
 do_install() {
-	oe_runmake install DESTDIR=${D} PREFIX=${prefix}
+	if test -e CMakeLists.txt ; then
+		cmake_do_install
+	else
+		oe_runmake install DESTDIR=${D} PREFIX=${prefix}
+	fi
 }
